+++ src/scorer/feature/image_attachment_scorer.ts
@@ -12,6 +12,6 @@ export default class ImageAttachmentScorer extends FeatureScorer {
     }
 
     async _score(toot: Toot) {
-        return (toot.reblog || toot).imageAttachments.length;
+        return toot.realToot().imageAttachments.length;
     }
 };