+++ src/scorer/feature/trending_tags_scorer.ts
@@ -15,7 +15,7 @@ export default class TrendingTagsScorer extends FeatureScorer {
     }
 
     async _score(toot: Toot) {
-        const tags = (toot.reblog || toot).trendingTags || [];
+        const tags = toot.realToot().trendingTags || [];
         const tagScores = tags.map(tag => tag.numAccounts || 0);
         let score = sumArray(tagScores);