+++ src/scorer/scorer.ts
@@ -55,7 +55,6 @@ export default abstract class Scorer {
         const rawScores = {} as StringNumberDict;
         const weightedScores = {} as StringNumberDict;
         const userWeights = await Storage.getWeightings();
-        const actualToot = toot.reblog ?? toot;
         const scores = await Promise.all(scorers.map((s) => s.score(toot)));
 
         // Compute a weighted score a toot based by multiplying the value of each numerical property
@@ -65,7 +64,7 @@ export default abstract class Scorer {
             rawScores[scorer.name] = scoreValue;
             weightedScores[scorer.name] = scoreValue * (userWeights[scorer.name] ?? 0);
 
-            if (actualToot.isTrending()) {
+            if (toot.realToot().isTrending()) {
                 weightedScores[scorer.name] *= (userWeights[WeightName.TRENDING] ?? 0);
             }
         });
@@ -77,7 +76,8 @@ export default abstract class Scorer {
         const weightedScore = this.sumScores(weightedScores);
 
         // Preserve rawScores, timeDecayMultiplier, and weightedScores for debugging
-        actualToot.scoreInfo = {
+        // TODO: duping the score to realToot() is a hack that sucks
+        toot.realToot().scoreInfo = toot.scoreInfo = {
             rawScore: this.sumScores(rawScores),
             rawScores,
             score: weightedScore * timeDecayMultiplier,
@@ -85,9 +85,6 @@ export default abstract class Scorer {
             weightedScores,
             weightedScore,
         } as TootScore;
-
-        // Copy the score info to the retoot if need be  // TODO: duping the score for retoots is a hack
-        toot.scoreInfo = actualToot.scoreInfo;
     }
 
     // Add 1 so that time decay multiplier works even with scorers giving 0s